Securing sql calls

+++ newUser.php
@@ -65,24 +65,20 @@
          if ($createUserBtn) {
             $userName = $userPassword = $confirmPassword = "";
 
-            if ($_SERVER["REQUEST_METHOD"] == "POST") {
+            if ($_SERVER["REQUEST_METHOD"] == "POST" && isset($_POST["userName"]) && isset($_POST["userPassword"]) && isset($_POST['confirmationPassword']) && isset($_POST['role'])) {
                $userName = $_POST["userName"];
                $userPassword = $_POST["userPassword"];
                $confirmationPassword = $_POST['confirmationPassword'];
 			   echo "<br/>[debug] post: ";
 			   print_r($_POST);
-               //$admin = $_POST["adminRole"];
-               //$user = $_POST["userRole"];
+			   
 			   $role = $_POST['role'];
 
                // Check if user exists
-               global $file_db;
-               $sql = "SELECT user_id FROM users WHERE user_name = '$userName'";
-               $result = $file_db->query($sql);
-               $result->setFetchMode(PDO::FETCH_ASSOC);
-               $result = $result->fetch();
-
-               if ($result['user_id']) {
+               $userId = getUserId($_POST["userName"]);
+			   
+				echo "<br/>[debug] userId verify: " . $userId;
+               if($userId != false) {
 
                   echo "<div class=\"container\"><div class=\"alert alert-warning\"><a href=\"#\" class=\"close\" data-dismiss=\"alert\" aria-label=\"close\">&times;</a><strong>Krap !</strong> This username is not available ! Choose another one ! </div></div>";
                } else {
@@ -93,28 +89,29 @@
                      // Check if password and confirmation password are identical
                      if ($userPassword == $confirmationPassword) {
 
-                        $userPasswordHash = password_hash($userPassword, PASSWORD_DEFAULT);
+                        $userPasswordHash = password_hash(htmlspecialchars($userPassword), PASSWORD_DEFAULT);
 
-                        /*if ($admin == 'role') {
+                        if ($role == 'Admin') {
                            $role = 1;
                         } else {
                            $role = 0;
-                        }*/
-
-                        $sql = "INSERT INTO users (user_name, user_pwd_hash, user_role, user_active, user_deleted)
-                                                           VALUES ('$userName', '$userPasswordHash', '$role', '1', '0')";
-                        $result = $file_db->query($sql);
+                        }
+						
+						$result = newUser($userName, $userPasswordHash, $role, 1, 0);
 
                         echo "<h2>New user created successfully !</h2>";
                         header("Location: http://localhost/html/admin.php?msg=created");
                      } else {
-                        echo "<h2>Password and confirmation password must match !</h2>";
+                        echo "<div class=\"container\"><div class=\"alert alert-warning\"><a href=\"#\" class=\"close\" data-dismiss=\"alert\" aria-label=\"close\">&times;</a><strong>Krap !</strong>  Passwords don't match ! </div></div>";
                      }
                   } else {
                      echo "<h2>Password must contain at least one caracter !</h2>";
                   }
                }
             }
+			else{
+				echo "<div class=\"container\"><div class=\"alert alert-warning\"><a href=\"#\" class=\"close\" data-dismiss=\"alert\" aria-label=\"close\">&times;</a><strong>Krap !</strong> All fields must be filled ! </div></div>";
+			}
          }
       ?>
       <?php include("includes/footer.php"); ?>
